Enable two forms to coexist on same page without validation error messages colliding
I added an option to `BootForm::open()` named `errorBag`.  Its purpose is to match the `$errorBag` property on a `FormRequest`, so it will be possible to have two or more forms on the same page (with the same field names), and only display the error messages associated with the form that was actually submitted.

The best example of this is a screen where a user can Login or Register.  Both forms have `email` and `password` fields.  If you were to incorrectly submit the login form without a password, the `AuthController.postLogin` method would catch this, populate the `default` errorBag (as defined in `ViewMessageBag`), and redirect you back to the page, and display the error message next to the password field in the login form.  However, because you also have a Register form on the same HTML page that is also pulling its validation errors from the `default` session error bag, the "The password field is required" error message meant for the Login form erroneously displays next to the password field on the Register form too.

Having at least one of the Controllers (or FormRequest classes) push its validation errors into an errorBag with a name other than `default` eliminates this problem.

Here's an example from my Login/Register page:

In my custom `RegisterController.postRegister(Request $request)` method, I send the user back to the Login/Register page, but I populate an errorBag named `register` instead of the default `default`:
````
$validator = Validator::make($request->input(), $rules);
if ($validator->fails()) {
	return back()->withInput()->withErrors($validator, 'registerHCP');
}
````

The login form:
````
{!! BootForm::open(['route'=>'login.post']) !!}
{!! BootForm::text('email', 'Email address') !!}
{!! BootForm::password('password', 'Password') !!}
{!! BootForm::checkbox('remember', 'Remember me') !!}
{!! BootForm::submit() !!}
{!! BootForm::close() !!}
````

The register form on the same HTML page:
````
{!! BootForm::open(['route' => 'register.post', 'errorBag'=>'register']) !!}
{!! BootForm::text('name', 'First name') !!}
{!! BootForm::email('email', 'Email address') !!}
{!! BootForm::tel('phone', 'Your phone number') !!}
{!! BootForm::password('password', 'Password') !!}
{!! BootForm::submit() !!}
{!! BootForm::close() !!}
````

So now BootForm pulls the error messages for the Registration form from an error bag named `register`, and the Login page continues to pull from an error bag named `default`, so they do not conflict anymore.